Fix crash in create-element-to-jsx when using spread args (#80)
The default case for child nodes is to wrap them in `JSXExpressionContainer`
nodes, but that doesn't work when a spread arg is given as a child. Spread args
in `React.createElement` are actually unnecessary (and probably bad practice)
because React can handle an array given as a child. So to fix, I just got the
underlying value from the spread (generally an array) and use that as the child.

Note that this could potentially introduce key warnings because React now sees
an actual array, but those warnings are probably good.